+++ pdf/test/widget_flex_test.dart
@@ -75,6 +75,24 @@ void main() {
     );
   });
 
+  test('Flex Widgets Spacer', () {
+    pdf.addPage(
+      Page(
+        build: (Context context) => Column(
+          children: <Widget>[
+            Text('Begin'),
+            Spacer(), // Defaults to a flex of one.
+            Text('Middle'),
+            // Gives twice the space between Middle and End than Begin and Middle.
+            Spacer(flex: 2),
+            // Expanded(flex: 2, child: SizedBox.shrink()),
+            Text('End'),
+          ],
+        ),
+      ),
+    );
+  });
+
   tearDownAll(() {
     final File file = File('widgets-flex.pdf');
     file.writeAsBytesSync(pdf.save());