+++ pdf/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix PdfColors.fromHex()
 - Update Barcode library to 1.9.0
 - Fix exif orientation crash
+- Fix Spacer Widget
 
 ## 1.7.1